+++ http_gen/CHANGELOG.md
@@ -1,3 +1,9 @@
+### 0.0.6
+
+- support toJson from mixins (eg freezed)
+- use compute function on flutter if HttpApi.useFlutterCompute is ture
+- fix not throwing the response body if toJson is not found
+
 ### 0.0.4
 
 move example to http_annotations